+++ integration/app/app.state.ts
@@ -1,0 +1,7 @@
+import { TodoStateModel, TodosState, TodoState } from './todo.state';
+
+export interface AppState {
+  todos: TodoStateModel;
+}
+
+export const states = [TodosState, TodoState];