+++ integration/app/list/list.module.ts
@@ -3,11 +3,11 @@ import { ListComponent } from './list.component';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { NgxsModule } from 'ngxs';
-import { ListStore } from './list.store';
+import { ListState } from './list.state';
 import { routes } from './list.routes';
 
 @NgModule({
-  imports: [CommonModule, RouterModule.forChild(routes), NgxsModule.forFeature([ListStore])],
+  imports: [CommonModule, RouterModule.forChild(routes), NgxsModule.forFeature([ListState])],
   declarations: [ListComponent]
 })
 export class ListModule {}