+++ integration/app/list/list.store.ts
@@ -1,7 +1,0 @@
-import { Store } from 'ngxs';
-
-@Store({
-  name: 'list',
-  defaults: ['foo']
-})
-export class ListStore {}