+++ src/index.ts
@@ -1,18 +1,12 @@
 export { NgxsModule } from './module';
 export { Action } from './action';
-export { Mutation } from './mutation';
-export { Ngxs } from './ngxs';
 export { Store } from './store';
+export { State } from './state';
 export { Select } from './select';
 export { EventStream } from './event-stream';
-export { ofEvent } from './of-event';
-export { NgxsPlugin, NgxsPluginFn } from './symbols';
-export {
-  ReduxDevtoolsPlugin,
-  LoggerPlugin,
-  LocalStoragePlugin,
-  LocalStoragePluginModule,
-  StorageStrategy,
-  LoggerPluginModule,
-  ReduxDevtoolsPluginModule
-} from './plugins';
+export { ofAction } from './of-action';
+export { NgxsPlugin, NgxsPluginFn, StateContext } from './symbols';
+
+export { ReduxDevtoolsPlugin, ReduxDevtoolsPluginModule } from './plugins/devtools/index';
+export { LoggerPlugin, LoggerPluginModule } from './plugins/logger/index';
+export { LocalStoragePlugin, LocalStoragePluginModule } from './plugins/localstorage/index';