+++ src/plugins/localstorage/localstorage.plugin.ts
@@ -1,8 +1,8 @@
 import { Inject, Injectable } from '@angular/core';
 import { NgxsPlugin } from '../../symbols';
-import { LocalStoragePluginOptions, LOCAL_STORAGE_PLUGIN_OPTIONS, StorageStrategy } from './symbols';
-import { getTypeFromInstance } from '../../internals';
-import { setValue, getValue } from './utils';
+import { LocalStoragePluginOptions, LOCAL_STORAGE_PLUGIN_OPTIONS } from './symbols';
+import { getTypeFromInstance, setValue, getValue } from '../../internals';
+import { tap } from 'rxjs/operators';
 
 @Injectable()
 export class LocalStoragePlugin implements NgxsPlugin {
@@ -12,7 +12,7 @@ export class LocalStoragePlugin implements NgxsPlugin {
     const options = this._options || <any>{};
     const isInitAction = getTypeFromInstance(event) === '@@INIT';
     const keys = Array.isArray(options.key) ? options.key : [options.key];
-    const engine = options.strategy === StorageStrategy.localstorage ? localStorage : sessionStorage;
+    const engine = !options.storage ? localStorage : options.storage;
 
     if (isInitAction) {
       for (const key of keys) {
@@ -30,20 +30,20 @@ export class LocalStoragePlugin implements NgxsPlugin {
       }
     }
 
-    state = next(state, event);
+    return next(state, event).pipe(
+      tap(nextState => {
+        if (!isInitAction) {
+          for (const key of keys) {
+            let val = nextState;
 
-    if (!isInitAction) {
-      for (const key of keys) {
-        let val = state;
+            if (key !== '@@STATE') {
+              val = getValue(nextState, key);
+            }
 
-        if (key !== '@@STATE') {
-          val = getValue(state, key);
+            engine.setItem(key, options.serialize(val));
+          }
         }
-
-        engine.setItem(key, options.serialize(val));
-      }
-    }
-
-    return state;
+      })
+    );
   }
 }