+++ src/plugins/logger/index.ts
@@ -1,0 +1,2 @@
+export { LoggerPluginModule } from './logger.module';
+export { LoggerPlugin } from './logger.plugin';