+++ src/plugins/logger/logger.module.ts
@@ -1,0 +1,30 @@
+import { NgModule, ModuleWithProviders } from '@angular/core';
+import { LoggerPluginOptions, LOGGER_PLUGIN_OPTIONS } from './symbols';
+import { NGXS_PLUGINS } from '../../symbols';
+import { LoggerPlugin } from './logger.plugin';
+
+@NgModule()
+export class LoggerPluginModule {
+  static forRoot(options: LoggerPluginOptions): ModuleWithProviders {
+    return {
+      ngModule: LoggerPluginModule,
+      providers: [
+        {
+          provide: NGXS_PLUGINS,
+          useClass: LoggerPlugin,
+          multi: true
+        },
+        {
+          provide: LOGGER_PLUGIN_OPTIONS,
+          useValue: {
+            ...{
+              logger: console,
+              collapsed: false
+            },
+            ...options
+          }
+        }
+      ]
+    };
+  }
+}